+++ bcitflex/app_functions/course_query.py
@@ -1,40 +1,88 @@
 """Filter courses."""
-from sqlalchemy import select
-from sqlalchemy.orm import Session
+from typing import Type
 
-from bcitflex.model import Course
+from functional import seq
+from sqlalchemy import Column, inspect
+
+from bcitflex.model import Base, Course, Subject
 
 
-def filter_courses(
-    session: Session,
-    subject: str | None = None,
-    course_code: str | None = None,
-    available: bool | None = None,
-    name: str | None = None,
-) -> list[Course]:
-    """Return a list of courses that match the given criteria."""
-    # Create a list of filters to apply to the query
-    joins = []
-    filters = []
+def coerce_to_column_type(model_column: Column, value: str | int) -> str | int:
+    """Coerce a value to the type of the given model column."""
+    column_data_type = model_column.type.python_type
+    return column_data_type(value)
 
-    if subject:
-        filters.append(Course.subject_id == subject)
-    if course_code:
-        filters.append(Course.code == course_code)
-    if name:
-        filters.append(Course.name.ilike(f"%{name}%"))
 
-    # Apply the filters to the query
-    query = select(Course)
-    for join in joins:
-        query.join(join)
-    query = query.where(*filters)
+class ModelFilter:
+    def __init__(self, model: Type[Base]):
+        self.conditions: list[callable] = []
+        self.model = model
+        self.mapper = inspect(model)
+        self.relationships = {
+            relationship.mapper.class_: relationship.key
+            for relationship in self.mapper.relationships
+        }
 
-    # Execute the query
-    courses = session.scalars(query).all()
+    def __call__(self, obj: Type[Base]) -> bool:
+        return all(condition(obj) for condition in self.conditions)
 
-    # Apply object filters
-    if available:
-        courses = [course for course in courses if course.is_available]
+    def __repr__(self) -> str:
+        return f"{self.__class__.__name__}({self.conditions})"
 
-    return courses
+    def add_condition(
+        self, attr: str, value: str | int | None, relation: Type[Base] | None = None
+    ):
+        """Return a function that returns True if the given model attribute
+        matches the given value.
+        """
+
+        if value is None or value == "":
+            return
+
+        attr_model = relation or self.model
+        attr_model_mapper = inspect(attr_model)
+        if attr in attr_model_mapper.columns:
+            model_column = attr_model_mapper.columns[attr]
+            value = coerce_to_column_type(model_column, value)
+
+        if relation is not None:
+            rel_key = self.relationships[relation]
+
+            if self.mapper.relationships[rel_key].uselist:
+
+                def condition(obj: Type[Base]) -> bool:
+                    model_attr = getattr(obj, rel_key)
+                    return any(getattr(item, attr) == value for item in model_attr)
+
+            else:
+
+                def condition(obj: Type[Base]) -> bool:
+                    model_attr = getattr(getattr(obj, rel_key), attr)
+                    return model_attr == value
+
+        else:
+
+            def condition(obj: Type[Base]) -> bool:
+                model_attr = getattr(obj, attr)
+                return model_attr == value
+
+        self.conditions.append(condition)
+
+    def filter(self, objs: list[Base]) -> list[Base]:
+        """Return a list of objects that match the given criteria."""
+        objs_seq = seq(objs)
+        objs_seq = objs_seq.filter(self)
+        return objs_seq.to_list()
+
+
+if __name__ == "__main__":
+    courses = [
+        Course(code="1234", subject=Subject(subject_id="COMP")),
+        Course(code="5678", subject=Subject(subject_id="COMP")),
+    ]
+
+    course_filter = ModelFilter(Course)
+    course_filter.add_condition("subject_id", "COMP", Subject)
+    course_filter.add_condition("code", "1234")
+    filtered_courses = course_filter.filter(courses)
+    print(filtered_courses)