+++ tests/app/test_course.py
@@ -1,37 +1,35 @@
 """Test the course blueprint."""
-from bs4 import BeautifulSoup
-
 from tests import dbtest
 
 
 @dbtest
-def test_index(client, auth):
-    response = client.get("/")
-    assert b"Log In" in response.data
-    assert b"Register" in response.data
+class TestCourse:
+    def test_auth(self, client, auth):
+        """Test that the response changes based on authentication."""
+        response = client.get("/")
+        assert b"Log In" in response.data
+        assert b"Register" in response.data
 
-    auth.login()
-    response = client.get("/")
-    assert b"test-user" in response.data
+        auth.login()
+        response = client.get("/")
+        assert b"test-user" in response.data
 
+    def test_index(self, client):
+        """Test that index page loads list of courses."""
+        response = client.get("/")
+        assert b"COMP" in response.data
 
-@dbtest
-def test_courses_name_filter(client, auth):
-    data = {"name": "Test"}
-    response = client.post("/courses", data=data)
-    soup = BeautifulSoup(response.data, "html.parser")
-    number_rows = len(soup.find_all("tr", id="course_data"))
-    assert b"Test Course" in response.data
-    assert b"Second Course" not in response.data
-    assert number_rows == 1
+    def test_filter_courses(self, client):
+        """Test that filter courses returns only courses with the given subject."""
+        response = client.post("/courses", data={"subject": "COMP"})
+        assert b"COMP" in response.data
 
+    def test_filter_courses_by_code(self, client):
+        """Test that filter courses returns only courses with the given code."""
+        response = client.post("/courses", data={"code": 9999})
+        assert b"1234" not in response.data
 
-@dbtest
-def test_courses_with_cours(client, auth):
-    data = {"name": "cours"}
-    response = client.post("/courses", data=data)
-    soup = BeautifulSoup(response.data, "html.parser")
-    number_rows = len(soup.find_all("tr", id="course_data"))
-    assert b"Test Course" in response.data
-    assert b"Second Course" in response.data
-    assert number_rows == 2
+    def test_filter_courses_by_available(self, client):
+        """Test that filter courses returns only courses with the given availability."""
+        response = client.post("/courses", data={"available": "True"})
+        assert b"COMP" in response.data